feat(core): add `path` property to modules
Some apps may quickly become very large, and the current nest architecture does not allow much room
to scale. This commit adds a `path` property on module metadata. This way, one can maintain good
modular structure without worrying about maintaining the right nested paths. It takes advantage of
the path package in Node.JS to concatenate paths without worrying about slashes (this may need to be
changed or standardized in the future, by checking for slashes in the first or last position of the
string). This is a MINOR VERSION change, and should not be merged until the next minor is about to
release.

Fixes #255