fix(core): Make nest compatible with fuse-box bundler
In order to package a nest application into a binary, tools like nexe use a bundler. Unfortunately bundlers don't always interpret the 'this' keyword appropriately. This removes the use of 'this' to avoid confusing a bundle, and adds assertions to the existing test suite to ensure the behavior remains the same.