Bugfix: flattening array from a route in interceptors
GIVEN: A route that returns an array as a promise.
WHEN: You are using an interceptor (even empty) that executes next.handle() and do
      nothing more. (or does something, it doesn't matter)
ACTUAL: You have an RxJS message for every single item in the array returned
        by the route. And after the last message (last item from the array)
        the observable completed and the only last item is returned from the
        API endpoint instead of the array of all these items.
EXPECTED: InterceptorsConsumer class should not flatten arrays that routes
          returns to separate RxJS messages. It should one message. And as the
          result API endpoint should return the whole array of items

This commit changes the behavior to the expected.

Why have it happened?
It happened because switchAll()/mergeAll() RxJS operators flat arrays if the
message is an array. However, if the message wrapped by a Promise or an
Observable, so these operators doesn't do the flattening