fix(core): convert URIError thrown by the HTTP platform to BadRequestException
When given a path parameter with an invalid percent encoding, Express throws a URIError with the
statusCode field set to 400 (bad request):
https://github.com/expressjs/express/blob/4.17.1/lib/router/layer.js#L166

Convert this URIError to a BadRequestException before providing it to the exception filters.
Therefore the exception filters will handle this error in the same way as other bad requests.

Note that this only applies to errors thrown directly from Express or Fastify. A URIError thrown
from a NestJS construct such as a controller or validation pipe will continue to be provided to the
exception filters as is.

Closes #8915