test: split abrupt stop test into a utils file
As we make use of the `http` module instead of axios or supertest
we have tomake use of some rather low level code. Using this utils
file we can have clearer names of what is happening in each request
to make it easier to follow what's happening. I've also added in some
comments about why each part is the way it is for clarity.