test: update the test for cache functionality
In cache-manager@^4, the `ttl` option was a number that represented the
number of seconds. In cache-manager@^5 the `ttl` is now a number that
represents the number of milliseconds. This difference caused our tests
to be flaky. I've also updated the default in the case of v5 to be 5000
instead of just 5 so that we get the same functionality for the default
value.