fix(common): correct order for logger error parameters
The error() function of a logger has two optional parameters: the trace and the context. The logger
now no longer passes the global context as the first optional parameter when there are is no value
supplied for the trace parameter. Instead, the trace parameter is passed as undefined and the
context is passed as the second optional parameter.

This fixes a bug where the trace is overwritten by the context when passed to custom loggers.
Fixes gremo/nest-winston#473