perf(common): In case of an expensive log, allow to pass a function
In certain case, it is very expensive to log some values. Instead of
checking if certain level is enabled, allow the users to pass in a
function that is only called if we need to stringify the log value.

This keeps the API small, while still allowing expensive logging to
be called when needed, and skipped when not.

e.g. instead of

```javascript
// This check is technically done twice.
if (this.logger.isLevelEnabled('debug')) {
  this.logger.debug(this.generateAsciiArtFromJpeg());
}
```

the customer could use:

```javascript
this.logger.debug(() => this.generateAsciiArtFromJpeg());
```