test(common): Fix invalid file type test
Fixes a unit test, which aims to provide a PNG instead of JPEG.
The unit test passes because the file type cannot be recognized,
yet the intention was more likely to detect it was a different file type.

Change the PNG data into valid PNG file data.
Ensure the PNG data is detected as PNG.