feat(common): add pipes for multiple files validation
When dealing with file uploads, NestJS offers the `ParseFilePipe` to help validate files uploaded under the same name. But when several files are uploaded with different names, there is not built-in pipe available and the validation is usually done in the controller method or service. This commit add two new pipes, `ParseFileFieldsPipe` and `GroupFilesPipe` that can be used two validate multiple files together in association with the `UploadedFiles` decorator and `FileValidator`.