feat: remove use of pipeline
Pipeline is great for when multiple streams need to be used together
but can also end up causing problems when there's a readstream of a
large file being sent in chunks, because the streams get destroyed
early. Technically, everything works fine, but there's warnings in
the terminal that could be misleading, so calling the streams with
error handlers directly is a safer way to go about it.