fix: prevent stack overflow with snapshot enabled
The most lazy fix I could think of to prevent the stack overflow during search for the next id. It will not help with the ever-increasing registry size and slowness with snapshot enabled in the application until you grab the next snapshot, but it will prevent the stack overflow if the search goes too deep.

Feel free to fix it another way if you're dissatisfied with this band-aid. Just wanted to highlight the issue.